Use enable_unstable build flag to control inclusion of persistent buffering capability
