+++ exporter/exporterhelper/persistent_queue_test.go
@@ -12,6 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+//go:build enable_unstable
+// +build enable_unstable
+
 package exporterhelper
 
 import (
